chore: move `trace.cpp` to kernel
Motivation: trace kernel `is_def_eq`
