+++ src/kernel/trace.h
