+++ src/library/compiler/compiler.cpp
@@ -8,8 +8,8 @@ Author: Leonardo de Moura
 #include "util/io.h"
 #include "kernel/type_checker.h"
 #include "kernel/kernel_exception.h"
+#include "kernel/trace.h"
 #include "library/max_sharing.h"
-#include "library/trace.h"
 #include "library/time_task.h"
 #include "library/compiler/util.h"
 #include "library/compiler/lcnf.h"
