+++ src/library/compiler/csimp.cpp
@@ -15,10 +15,10 @@ Author: Leonardo de Moura
 #include "kernel/instantiate.h"
 #include "kernel/inductive.h"
 #include "kernel/kernel_exception.h"
+#include "kernel/trace.h"
 #include "library/util.h"
 #include "library/constants.h"
 #include "library/class.h"
-#include "library/trace.h"
 #include "library/expr_pair_maps.h"
 #include "library/compiler/util.h"
 #include "library/compiler/cse.h"
