+++ src/library/compiler/ir_interpreter.cpp
@@ -40,8 +40,8 @@ functions, which have a (relatively) homogeneous ABI that we can use without run
 #include "runtime/io.h"
 #include "runtime/option_ref.h"
 #include "runtime/array_ref.h"
+#include "kernel/trace.h"
 #include "library/time_task.h"
-#include "library/trace.h"
 #include "library/compiler/ir.h"
 #include "library/compiler/init_attribute.h"
 #include "util/nat.h"
