+++ src/library/compiler/lambda_lifting.cpp
@@ -10,7 +10,7 @@ Author: Leonardo de Moura
 #include "kernel/abstract.h"
 #include "kernel/for_each_fn.h"
 #include "kernel/inductive.h"
-#include "library/trace.h"
+#include "kernel/trace.h"
 #include "library/compiler/util.h"
 #include "library/compiler/closed_term_cache.h"
