+++ src/library/compiler/llnf.cpp
@@ -16,8 +16,8 @@ Author: Leonardo de Moura
 #include "kernel/abstract.h"
 #include "kernel/for_each_fn.h"
 #include "kernel/inductive.h"
+#include "kernel/trace.h"
 #include "library/util.h"
-#include "library/trace.h"
 #include "library/compiler/util.h"
 #include "library/compiler/llnf.h"
 #include "library/compiler/ll_infer_type.h"
