+++ src/library/compiler/util.cpp
@@ -18,12 +18,12 @@ Author: Leonardo de Moura
 #include "kernel/inductive.h"
 #include "kernel/instantiate.h"
 #include "kernel/kernel_exception.h"
+#include "kernel/trace.h"
 #include "library/util.h"
 #include "library/suffixes.h"
 #include "library/aux_recursors.h"
 #include "library/replace_visitor.h"
 #include "library/constants.h"
-#include "library/trace.h"
 #include "library/compiler/lambda_lifting.h"
 #include "library/compiler/eager_lambda_lifting.h"
 #include "library/compiler/util.h"
