+++ src/library/init_module.cpp
@@ -4,7 +4,6 @@ Released under Apache 2.0 license as described in the file LICENSE.
 
 Author: Leonardo de Moura
 */
-#include "library/trace.h"
 #include "library/constants.h"
 #include "library/class.h"
 #include "library/num.h"
@@ -21,11 +20,9 @@ void initialize_library_core_module() {
     initialize_formatter();
     initialize_constants();
     initialize_profiling();
-    initialize_trace();
 }
 
 void finalize_library_core_module() {
-    finalize_trace();
     finalize_profiling();
     finalize_constants();
     finalize_formatter();
