+++ src/util/shell.cpp
@@ -29,11 +29,11 @@ Author: Leonardo de Moura
 #include "util/option_declarations.h"
 #include "kernel/environment.h"
 #include "kernel/kernel_exception.h"
+#include "kernel/trace.h"
 #include "library/formatter.h"
 #include "library/module.h"
 #include "library/time_task.h"
 #include "library/compiler/ir.h"
-#include "library/trace.h"
 #include "library/print.h"
 #include "initialize/init.h"
 #include "library/compiler/ir_interpreter.h"
